+++ src/crypto/index.ts
@@ -1,6 +1,6 @@
 export * from './core';
 export * from './NUT01';
-export * from './NUT09';
 export * from './NUT11';
+export * from './NUT13';
 export * from './NUT12';
 export * from './NUT20';
